backend/requirements.txt
django==5.0
djangorestframework
django-cors-headers
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
whitenoise==6.6.0
python-dotenv
djangorestframework-simplejwt